+++ src/briefy/choreographer/docs/__init__.py
@@ -1,0 +1,1 @@
+"""Sphinx helpers to briefy.choreographer."""