+++ src/briefy/choreographer/subscribers/leica/task/__init__.py
@@ -1,0 +1,1 @@
+"""Leica subscribers for Leica Task events."""